final fix Icon folder rename
