+++ src/components/Icon/IconNames.ts
