+++ src/components/styledComponents/FilterTag.tsx
@@ -1,6 +1,6 @@
 import { MouseEventHandler, PropsWithChildren, useCallback } from "react";
 import styled from "styled-components";
-import Icon from "../IconComponent";
+import Icon from "../Icon";
 
 interface Props {
   selected?: boolean;
